services/telegram-bot/src/main.py
import logging
from telegram import Update
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes
import grpc
import proto.service_pb2 as service_pb2
import proto.service_pb2_grpc as service_pb2_grpc

TELEGRAM_TOKEN = "your-telegram-bot-token"
API_GATEWAY_ADDRESS = "api-gateway:50051"

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("Send me an audio or video file for transcription")

async def handle_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("Not implemented")

def main():
    logging.basicConfig(level=logging.INFO)
    app = Application.builder().token(TELEGRAM_TOKEN).build()
    app.add_handler(CommandHandler("start", start))
    app.add_handler(MessageHandler(filters.ALL, handle_message))
    app.run_polling()

if __name__ == "__main__":
    main()
